modals implemented

+++ app/categories/page.jsx
@@ -6,8 +6,10 @@ import axios from 'axios'
 import { getProviders, useSession, signIn, signOut } from 'next-auth/react'
 import Image from "next/image"
 import logo from "@public/logo.png"
+import { toast } from 'react-hot-toast'
+import { withSwal } from 'react-sweetalert2'
 
-const Page = () => {
+const Page = ({ swal }) => {
 
     const { data: session } = useSession();
 
@@ -29,6 +31,11 @@ const Page = () => {
 
     const createCategory = async () => {
 
+        if (!categoryName) {
+            toast.error("Category name cannot be null")
+            return
+        }
+
         if (!editData) {
             await axios.post("/api/category/new", {
                 categoryName,
@@ -40,11 +47,18 @@ const Page = () => {
             }).then((response) => {
                 setCategoryName("")
                 setParentCategory("")
+                toast.success(`${categoryName} added`)
             }).catch((err) => {
                 console.log(err.message)
             })
         }
         else {
+
+            if (!categoryName || !parentCategory) {
+                toast.error("Category name cannot be null")
+                return
+            }
+
             await axios.put(`/api/category/edit`, {
                 categoryName,
                 parentCategory,
@@ -76,12 +90,29 @@ const Page = () => {
     }
 
     const deleteCategory = async (category) => {
-        await axios.delete(`/api/category/delete/${category._id}`).then((response) => {
-            console.log(response)
-        }).catch((err) => {
-            console.log(err.message)
-        })
-        getCategories()
+
+        swal.fire({
+            title: 'Are You Sure?',
+            text: 'Confirm delete ' + category.categoryName,
+            showCancelButton: true,
+            cancelButtonText: 'Cancel',
+            confirmButtonText: 'Yes, Delete!',
+            reverseButtons: true,
+            confirmButtonColor: '#d55',
+
+        }).then(async result => {
+            if (result.isConfirmed) {
+                await axios.delete(`/api/category/delete/${category._id}`).then((response) => {
+                    // console.log(response)
+                    toast.success(`${category.categoryName} deleted`)
+                }).catch((err) => {
+                    console.log(err.message)
+                })
+                getCategories()
+            }
+        }).catch(error => {
+            toast.error(err.message)
+        });
     }
 
     const getCategories = async () => {
@@ -132,7 +163,10 @@ const Page = () => {
             <div className="flex flex-col items-center justify-center">
                 <Image src={logo} width={250} height={250} alt="Company logo" />
                 <span className="text-white font-bold text-lg">Welcome to Shop-it Admin Portal</span>
-                <button key={providers?.name} onClick={() => signIn('google')} className="bg-white h-[2.5rem] w-[12rem] mt-5 rounded-lg text-blue-900 font-bold flex items-center justify-center gap-2">
+                <button key={providers?.name} onClick={async () => {
+                    await signIn('google')
+                    toast.success("Logged In")
+                }} className="bg-white h-[2.5rem] w-[12rem] mt-5 rounded-lg text-blue-900 font-bold flex items-center justify-center gap-2">
                     <span>
                         <svg className="w-4 h-4" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 488 512"><path d="M488 261.8C488 403.3 391.1 504 248 504 110.8 504 0 393.2 0 256S110.8 8 248 8c66.8 0 123 24.5 166.3 64.9l-67.5 64.9C258.5 52.6 94.3 116.6 94.3 256c0 86.5 69.1 156.6 153.7 156.6 98.2 0 135-70.4 140.8-106.9H248v-85.3h236.1c2.3 12.7 3.9 24.9 3.9 41.4z" /></svg>
                     </span>
@@ -211,12 +245,22 @@ const Page = () => {
                                     )
                                 })
                                 }
-                                <button onClick={addProperty} className='w-[200px] bg-blue-900 text-white font-bold h-10 rounded-xl hover:border hover:border-blue-900 hover:text-blue-900 hover:bg-white transition-all duration-300'>Add Properties</button>
+                                <button onClick={addProperty} className='w-[200px] bg-blue-900 text-white font-bold h-10 rounded-xl hover:border hover:border-blue-900 hover:text-blue-900 hover:bg-white transition-all duration-300 flex items-center justify-center gap-3'>
+                                    <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" className="w-6 h-6">
+                                        <path strokeLinecap="round" strokeLinejoin="round" d="M12 9v6m3-3H9m12 0a9 9 0 11-18 0 9 9 0 0118 0z" />
+                                    </svg>
+
+                                    Add Properties</button>
 
                             </div>
                         </div>
                         <div>{!editData ? (
-                            <button className='mt-2 mb-2 xl:mb-0 xl:mt-0 w-[200px] bg-blue-900 text-white font-bold h-10 rounded-xl hover:border hover:border-blue-900 hover:text-blue-900 hover:bg-white transition-all duration-300' onClick={createCategory}>Create</button>
+                            <button className='mt-2 mb-2 xl:mb-0 xl:mt-0 w-[200px] bg-blue-900 text-white font-bold h-10 rounded-xl hover:border hover:border-blue-900 hover:text-blue-900 hover:bg-white transition-all duration-300 flex items-center justify-center gap-3' onClick={createCategory}>
+                                <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" className="w-6 h-6">
+                                    <path strokeLinecap="round" strokeLinejoin="round" d="M12 4.5v15m7.5-7.5h-15" />
+                                </svg>
+
+                                Create</button>
 
                         ) : (
                             <div>
@@ -286,4 +330,8 @@ const Page = () => {
     }
 }
 
-export default Page
+// export default Page
+
+export default withSwal(({ swal }, ref) => (
+    <Page swal={swal} />
+))
