+++ app/utilities/page.jsx
@@ -6,8 +6,10 @@ import axios from 'axios'
 import { getProviders, useSession, signIn, signOut } from 'next-auth/react'
 import Image from "next/image"
 import logo from "@public/logo.png"
+import { toast } from 'react-hot-toast'
+import { withSwal } from 'react-sweetalert2'
 
-const page = () => {
+const Page = ({ swal }) => {
 
     const { data: session } = useSession();
 
@@ -24,24 +26,67 @@ const page = () => {
     }, [])
 
     const handleCategoryClear = async () => {
-        await axios.delete("/api/utilities/category").then((response) => {
-        }).catch((err) => {
-            console.log(err.message)
-        })
+
+        swal.fire({
+            title: 'Are You Sure?',
+            text: 'This is a destructive action, Proceed?',
+            showCancelButton: true,
+            cancelButtonText: 'Cancel',
+            confirmButtonText: 'Yes, Proceed!',
+            reverseButtons: true,
+            confirmButtonColor: '#d55',
+
+        }).then(async result => {
+            if (result.isConfirmed) {
+                await axios.delete("/api/utilities/category").then((response) => {
+                    toast.success("Category Database cleared")
+                }).catch((err) => {
+                    console.log(err.message)
+                })
+            }
+        }).catch(error => {
+            toast.error(err.message)
+        });
+
+
     }
 
     const handleProductClear = async () => {
-        await axios.delete("/api/utilities/products").then((response) => {
-        }).catch((err) => {
-            console.log(err.message)
-        })
+
+        swal.fire({
+            title: 'Are You Sure?',
+            text: 'This is a destructive action, Proceed?',
+            showCancelButton: true,
+            cancelButtonText: 'Cancel',
+            confirmButtonText: 'Yes, Proceed!',
+            reverseButtons: true,
+            confirmButtonColor: '#d55',
+
+        }).then(async result => {
+            if (result.isConfirmed) {
+                await axios.delete("/api/utilities/products").then((response) => {
+                    toast.success("Product Database cleared")
+                }).catch((err) => {
+                    console.log(err.message)
+                })
+            }
+        }).catch(error => {
+            toast.error(err.message)
+        });
     }
 
     const handleAddAdmin = async () => {
+
+        if (!admin) {
+            toast.error("Admin e-mail cannot be empty")
+            return
+        }
+
         await axios.post("/api/utilities/admin/add", {
             email: admin
         }).then((response) => {
             setAdmin("")
+            toast.success("Admin added")
         }).catch((err) => {
             console.log(err.message)
         })
@@ -57,12 +102,32 @@ const page = () => {
     }
 
     const handleAdminDelete = async (id) => {
-        await axios.delete(`/api/utilities/admin/delete/${id}`)
-            .then((response) => {
-                getAdmins()
-            }).catch((err) => {
-                console.log(err.message)
-            })
+
+
+        swal.fire({
+            title: 'Are You Sure?',
+            text: 'Confirm delete ' + admin,
+            showCancelButton: true,
+            cancelButtonText: 'Cancel',
+            confirmButtonText: 'Yes, Delete!',
+            reverseButtons: true,
+            confirmButtonColor: '#d55',
+
+        }).then(async result => {
+            if (result.isConfirmed) {
+                await axios.delete(`/api/utilities/admin/delete/${id}`)
+                    .then((response) => {
+                        getAdmins()
+                        toast.success("Admin deleted")
+                    }).catch((err) => {
+                        console.log(err.message)
+                    })
+            }
+        }).catch(error => {
+            toast.error(err.message)
+        });
+
+
     }
 
     useEffect(() => {
@@ -74,7 +139,10 @@ const page = () => {
             <div className="flex flex-col items-center justify-center">
                 <Image src={logo} width={250} height={250} alt="Company logo" />
                 <span className="text-white font-bold text-lg">Welcome to Shop-it Admin Portal</span>
-                <button key={providers?.name} onClick={() => signIn('google')} className="bg-white h-[2.5rem] w-[12rem] mt-5 rounded-lg text-blue-900 font-bold flex items-center justify-center gap-2">
+                <button key={providers?.name} onClick={async () => {
+                    await signIn('google')
+                    toast.success("Logged In")
+                }} className="bg-white h-[2.5rem] w-[12rem] mt-5 rounded-lg text-blue-900 font-bold flex items-center justify-center gap-2">
                     <span>
                         <svg className="w-4 h-4" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 488 512"><path d="M488 261.8C488 403.3 391.1 504 248 504 110.8 504 0 393.2 0 256S110.8 8 248 8c66.8 0 123 24.5 166.3 64.9l-67.5 64.9C258.5 52.6 94.3 116.6 94.3 256c0 86.5 69.1 156.6 153.7 156.6 98.2 0 135-70.4 140.8-106.9H248v-85.3h236.1c2.3 12.7 3.9 24.9 3.9 41.4z" /></svg>
                     </span>
@@ -99,7 +167,7 @@ const page = () => {
                     <div className="bg-white flex flex-col flex-grow ml-[-10px] m-2 rounded-xl p-10 gap-3 xl:gap-4 items-center overflow-hidden overflow-y-scroll">
                         <span className='text-blue-900 font-extrabold text-3xl mb-3'>Utilities</span>
                         <span className='text-blue-900 font-bold text-xl'>Clear All Products</span>
-                        <button onClick={handleProductClear} className='w-[200px] bg-blue-900 text-white font-bold h-10 rounded-xl hover:border hover:border-blue-900 hover:text-blue-900 hover:bg-white transition-all duration-300 flex justify-center items-center gap-4'>
+                        <button onClick={handleProductClear} className='w-[200px] bg-red-700 text-white font-bold h-10 rounded-xl hover:border hover:border-red-700 hover:text-red-700 hover:bg-white transition-all duration-300 flex justify-center items-center gap-4'>
                             <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" className="w-6 h-6">
                                 <path strokeLinecap="round" strokeLinejoin="round" d="M20.25 7.5l-.625 10.632a2.25 2.25 0 01-2.247 2.118H6.622a2.25 2.25 0 01-2.247-2.118L3.75 7.5m6 4.125l2.25 2.25m0 0l2.25 2.25M12 13.875l2.25-2.25M12 13.875l-2.25 2.25M3.375 7.5h17.25c.621 0 1.125-.504 1.125-1.125v-1.5c0-.621-.504-1.125-1.125-1.125H3.375c-.621 0-1.125.504-1.125 1.125v1.5c0 .621.504 1.125 1.125 1.125z" />
                             </svg>
@@ -109,7 +177,7 @@ const page = () => {
                         <span className='text-blue-900 font-bold text-xl'>
 
                             Clear All Categories </span>
-                        <button onClick={handleCategoryClear} className='w-[200px] bg-blue-900 text-white font-bold h-10 rounded-xl hover:border hover:border-blue-900 hover:text-blue-900 hover:bg-white transition-all duration-300 flex justify-center items-center gap-4'>
+                        <button onClick={handleCategoryClear} className='w-[200px] bg-red-700 text-white font-bold h-10 rounded-xl hover:border hover:border-red-700 hover:text-red-700 hover:bg-white transition-all duration-300 flex justify-center items-center gap-4'>
 
                             <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" className="w-6 h-6">
                                 <path strokeLinecap="round" strokeLinejoin="round" d="M20.25 7.5l-.625 10.632a2.25 2.25 0 01-2.247 2.118H6.622a2.25 2.25 0 01-2.247-2.118L3.75 7.5m6 4.125l2.25 2.25m0 0l2.25 2.25M12 13.875l2.25-2.25M12 13.875l-2.25 2.25M3.375 7.5h17.25c.621 0 1.125-.504 1.125-1.125v-1.5c0-.621-.504-1.125-1.125-1.125H3.375c-.621 0-1.125.504-1.125 1.125v1.5c0 .621.504 1.125 1.125 1.125z" />
@@ -153,7 +221,7 @@ const page = () => {
                         }
                         <input
                             type="email"
-                            placeholder='Enter e-mail'
+                            placeholder='Enter e-mail to be added as admin'
                             className='mr-2 lg:mr-0 xl:w-[350px] w-[230px] border border-gray-600 h-10 pl-3 rounded-xl focus:outline-blue-500'
                             value={admin}
                             onChange={e => setAdmin(e.target.value)}
@@ -171,4 +239,8 @@ const page = () => {
     }
 }
 
-export default page
+// export default Page
+
+export default withSwal(({ swal }, ref) => (
+    <Page swal={swal} />
+))
